Recodifica códigos de provincias o departamentos de Argentina

Función que permite recodificar código de provincia para utilizacion de grilla de Argentina y de departamentos para los 24 distritos subnacionales y hacer mas sencillala vinculacion con distinto tipo de bases de datos con diferentes indicadores.

recode_district(data, type = NULL)

Arguments

data
debe ser un 'data.frame' obtenido con get_grid. Las grillas grillas disponibles se pueden chequear con show_arg_grids.

type
cual es el código qye se quiere definir en la grilla. Las opciones son 'indra', 'indec' o 'iso'

Value

El valor por defecto es codprov y coddepto para provincia y departamentos, respectivamente. Estos corresponden a la codificacion de los escrutinios provisorios de elecciones nacionales. Se puede optar por la nomenclautra de indec *_censo para ambos niveles o ISO 3166-2, para el nivel provincial.

Details

Respecto los datos se puede consultar la documentación de ISO 3166-2 - International Organization for Standardization y del del INDEC. codprov es la codificacion con la que se publciaron las bases de datos de INDRA, empresa encargada por muchos años de la tarea y utilizados en la libería polAr.

Examples

get_grid("ARGENTINA")

#>    col row code                name
#> 1    1   1   10               JUJUY
#> 2    3   2   09             FORMOSA
#> 3    5   2   14            MISIONES
#> 4    1   2   17               SALTA
#> 5    2   2   23             TUCUMAN
#> 6    1   3   03           CATAMARCA
#> 7    3   3   06               CHACO
#> 8    4   3   05          CORRIENTES
#> 9    2   3   22 SANTIAGO DEL ESTERO
#> 10   2   4   04             CORDOBA
#> 11   4   4   08          ENTRE RIOS
#> 12   1   4   12            LA RIOJA
#> 13   3   4   21            SANTA FE
#> 14   3   5   02        BUENOS AIRES
#> 15   4   5   01                CABA
#> 16   1   5   18            SAN JUAN
#> 17   2   5   19            SAN LUIS
#> 18   2   6   11            LA PAMPA
#> 19   1   6   13             MENDOZA
#> 20   1   7   15             NEUQUEN
#> 21   2   7   16           RIO NEGRO
#> 22   1   8   07              CHUBUT
#> 23   1   9   20          SANTA CRUZ
#> 24   1  10   24    TIERRA DEL FUEGO

get_grid("ARGENTINA") %>%
   recode_district(type = "iso")

#>    col row code                name
#> 1    1   1 AR-C               JUJUY
#> 2    3   2 AR-B             FORMOSA
#> 3    5   2 AR-K            MISIONES
#> 4    1   2 AR-X               SALTA
#> 5    2   2 AR-W             TUCUMAN
#> 6    1   3 AR-H           CATAMARCA
#> 7    3   3 AR-U               CHACO
#> 8    4   3 AR-E          CORRIENTES
#> 9    2   3 AR-P SANTIAGO DEL ESTERO
#> 10   2   4 AR-Y             CORDOBA
#> 11   4   4 AR-L          ENTRE RIOS
#> 12   1   4 AR-F            LA RIOJA
#> 13   3   4 AR-M            SANTA FE
#> 14   3   5 AR-N        BUENOS AIRES
#> 15   4   5 AR-Q                CABA
#> 16   1   5 AR-R            SAN JUAN
#> 17   2   5 AR-A            SAN LUIS
#> 18   2   6 AR-J            LA PAMPA
#> 19   1   6 AR-D             MENDOZA
#> 20   1   7 AR-Z             NEUQUEN
#> 21   2   7 AR-S           RIO NEGRO
#> 22   1   8 AR-G              CHUBUT
#> 23   1   9 AR-T          SANTA CRUZ
#> 24   1  10 AR-V    TIERRA DEL FUEGO